Bugfix: 删除按钮崩溃问题修复及项目结构初步优化
修复了点击删除按钮时由于强制解包nil导致的崩溃问题。

同时，对项目结构进行了初步优化：
- 创建了Models, ViewModels, 和 Views文件夹。
- 将FileInfo, ConversionRecord, ConvertaViewModel, ContentView, DropZoneView, ConversionView, FileCard, FormatPicker, QualitySlider 等文件移动到了对应的新目录结构中。

在文件移动过程中，遇到了rename_file工具无法稳定工作的挑战，采用了 复制-创建-删除 的策略来完成文件迁移。
