+++ Converta/Models/FileInfo.swift
