+++ Converta/Views/QualitySlider.swift
